cleanup(rest): move code to matching `.cc` file
The `AsStatus()` functions were declared in `rest_response.h`, but
defined in `rest_response.cc`.  That seems unnecessarily confusing. The
test were also a bit inconsistent.

I also refactored a function that `clang-tidy` claims has high
complexity to make it less so.
