+++ google/cloud/google_cloud_cpp_rest_internal_unit_tests.bzl
@@ -23,7 +23,6 @@ google_cloud_cpp_rest_internal_unit_tests = [
     "internal/curl_http_payload_test.cc",
     "internal/curl_impl_test.cc",
     "internal/curl_rest_client_test.cc",
-    "internal/curl_rest_response_test.cc",
     "internal/curl_wrappers_test.cc",
     "internal/make_jwt_assertion_test.cc",
     "internal/oauth2_access_token_credentials_test.cc",
@@ -39,5 +38,6 @@ google_cloud_cpp_rest_internal_unit_tests = [
     "internal/openssl_util_test.cc",
     "internal/rest_parse_json_error_test.cc",
     "internal/rest_request_test.cc",
+    "internal/rest_response_test.cc",
     "internal/unified_rest_credentials_test.cc",
 ]
